requirements.txt
PyAudio~=0.2.13
speechkit~=2.1.1
google-cloud-speech==2.18.0
transformers~=4.26.1
openai==0.27.2
